[fix] (ethereum) reinstall test libs to devDep
