+++ package.json
@@ -68,6 +68,7 @@
     "eslint-plugin-jest": "^25.2.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^5.1.1",
+    "ganache": "^7.0.4",
     "husky": "^3.0.9",
     "jest": "^27.2.4",
     "jest-circus": "^27.2.4",
@@ -77,6 +78,7 @@
     "link-parent-bin": "^2.0.0",
     "lint-staged": "^11.2.6",
     "nock": "^11.7.1",
+    "node-fetch": "^2.6.7",
     "prettier": "^1.14.3",
     "rimraf": "^2.6.2",
     "rollup": "^2.27.0",
@@ -100,8 +102,6 @@
     "@bitaccess/coinlib-litecoin": "file:packages/coinlib-litecoin",
     "@bitaccess/coinlib-ripple": "file:packages/coinlib-ripple",
     "@bitaccess/coinlib-stellar": "file:packages/coinlib-stellar",
-    "@bitaccess/coinlib-tron": "file:packages/coinlib-tron",
-    "ganache": "^7.0.3",
-    "node-fetch": "^2.6.7"
+    "@bitaccess/coinlib-tron": "file:packages/coinlib-tron"
   }
 }
