+++ packages/coinlib-litecoin/src/@types/bs58check/index.d.ts
@@ -1,4 +1,4 @@
 declare module 'bs58check' {
-  import * as base58 from 'bs58'
+  import * as base58 from 'bs58check'
   export default base58
 }
